Removed all usage of typeid(T).hash_code

+++ src/engine/core/includes/core/IdGenerator.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <stdint.h>
 
-using Id = uint64_t;
+using Id = size_t;
 
 class IdGenerator {
 public:
